requirements.txt
flet
flet-route
matplotlib
pymongo
python-dotenv
bcrypt
